app/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
Markdown==3.3.3
mssql==1.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pymssql==2.1.5
pytz==2020.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
